Mirrormaker: split integration test to setup base class and base test class
